updated routes to not use protected
