+++ src/api/apiClient.ts
@@ -1,6 +1,10 @@
 import AppAxiosInstance from '../auth/axios';
 import { VolunteerLeaderboardItem } from '../containers/volunteerLeaderboard/ducks/types';
 import { TeamLeaderboardItem } from '../containers/teamLeaderboard/ducks/types';
+import {
+  BlockGeoData,
+  NeighborhoodGeoData,
+} from '../components/mapPageComponents/mapContainer/ducks/types';
 
 export interface ApiExtraArgs {
   readonly apiClient: ApiClient;
@@ -13,11 +17,15 @@ export interface ApiClient {
   readonly getTeamsLeaderboard: (
     previousDays: number | null,
   ) => Promise<TeamLeaderboardItem[]>;
+  readonly getBlockGeoData: () => Promise<BlockGeoData>;
+  readonly getNeighborhoodGeoData: () => Promise<NeighborhoodGeoData>;
 }
 
 enum ApiClientRoutes {
   USERS_LEADERBOARD = '/api/v1/leaderboard/users',
   TEAMS_LEADERBOARD = '/api/v1/leaderboard/teams',
+  GET_ALL_BLOCKS = 'api/v1/map/blocks',
+  GET_ALL_NEIGHBORHOODS = 'api/v1/map/neighborhoods',
 }
 
 const getUsersLeaderboard = (
@@ -36,9 +44,27 @@ const getTeamsLeaderboard = (
   }).then((response) => response.data.entries);
 };
 
+const getBlockGeoData = (): Promise<BlockGeoData> => {
+  return AppAxiosInstance.get(ApiClientRoutes.GET_ALL_BLOCKS).then(
+    (response) => {
+      return response.data;
+    },
+  );
+};
+
+const getNeighborhoodGeoData = (): Promise<NeighborhoodGeoData> => {
+  return AppAxiosInstance.get(ApiClientRoutes.GET_ALL_NEIGHBORHOODS).then(
+    (response) => {
+      return response.data;
+    },
+  );
+};
+
 const Client: ApiClient = Object.freeze({
   getUsersLeaderboard,
   getTeamsLeaderboard,
+  getBlockGeoData,
+  getNeighborhoodGeoData,
 });
 
 export default Client;
